Workflow_27_BlogWriter/Backend/requirements.txt
accelerate==0.28.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
asyncpg==0.31.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
controlnet_aux==0.0.10
cryptography==46.0.5
cycler==0.12.1
dataclasses-json==0.6.7
ddgs==9.10.0
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.27.2
distro==1.9.0
einops==0.8.2
executing==2.2.1
fake-useragent==2.2.0
fastapi==0.128.0
fastjsonschema==2.21.2
filelock==3.25.2
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-genai==1.68.0
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.3
importlib_metadata==9.0.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
json5==0.13.0
json_repair==0.58.6
jsonpatch==1.33
jsonpath-ng==1.8.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-ollama==1.0.1
langchain-openai==1.1.6
langchain-tavily==0.2.17
langchain-text-splitters==1.1.0
langgraph==1.0.9
langgraph-checkpoint==3.0.1
langgraph-checkpoint-postgres==3.0.4
langgraph-checkpoint-redis==0.3.5
langgraph-checkpoint-sqlite==3.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.1
langsmith==0.6.0
lark==1.3.1
lazy-loader==0.5
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook_shim==0.2.4
ntplib==0.4.0
numpy==2.4.0
ollama==0.6.1
openai==2.14.0
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
peft==0.10.0
pillow==12.1.0
platformdirs==4.5.1
primp==0.15.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
python-ulid==3.1.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.0
redisvl==0.14.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.1
Send2Trash==2.0.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlite-vec==0.1.6
stack-data==0.6.3
starlette==0.50.0
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
tifffile==2026.3.3
tiktoken==0.12.0
timm==1.0.26
tinycss2==1.4.0
tokenizers==0.15.2
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.39.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
watchdog==6.0.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
